# [MetaSchedule][Hexagon] Improve vectorization for standalone elementwise op